+++ tests/ServerTestCaseTest.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace luya\testsuite\tests;
+
+use luya\testsuite\cases\ServerTestCase;
+
+class ServerTestCaseTest extends ServerTestCase
+{
+    public function getConfigArray()
+    {
+        return [
+            'id' => 'servertest',
+            'basePath' => dirname(__DIR__),
+        ];
+    }
+}